Reorganize project: aesthetics
