+++ sharding/src/main/scala/zio/actors/sharding/Entity.scala
@@ -4,6 +4,8 @@ import com.devsisters.shardcake.EntityType
 
 trait Entity {
   type Msg
+
   def name: String
+
   val entityType: EntityType[Msg] = new EntityType[Msg](name) {}
 }
